+++ routes/auth/signup/organization/index.js
@@ -1,6 +1,7 @@
 var express = require('express');
 var router = express.Router();
 
-router.use('/signup', require('./signup'));
+router.use('/', require('./organization'));
+
 
 module.exports = router;
